Move files from fight package to crtp Ref: #2342
