+++ crtp/src/main/java/crtp/MmaBantamweightFighter.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package crtp.fight;
+package crtp;
 
 /**
  * MmaBantamweightFighter class.
